% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{impute_basics}
\alias{impute_basics}
\title{Impute using BASiCS}
\usage{
impute_basics(sce, ...)
}
\arguments{
\item{sce}{the SingleCellExperiment object}

\item{...}{additional arguments for BaSICS}
}
\value{
matrix of imputed expression values
}
\description{
Impute using BASiCS
}
