% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drimpute.R
\name{impute_drimpute}
\alias{impute_drimpute}
\title{Impute using DrImpute}
\usage{
impute_drimpute(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
matrix of imputed expression values
}
\description{
Impute using DrImpute
}
