% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_smooth.R
\name{impute_knn_smooth}
\alias{impute_knn_smooth}
\title{Impute using knn smooth}
\usage{
impute_knn_smooth(sce, k = ceiling(ncol(sce)/20))
}
\arguments{
\item{sce}{the SingleCellExperiment object}

\item{k}{the number of nearest neighbours to use for smoothing}
}
\value{
matrix of imputed expression values
}
\description{
Impute using knn smooth
}
