% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saver.R
\name{impute_saver}
\alias{impute_saver}
\title{Impute using SAVER}
\usage{
impute_saver(sce)
}
\arguments{
\item{sce}{the SingleCellExperiment object}
}
\value{
matrix of imputed expression values
}
\description{
Impute using SAVER
}
